// src/app/enums/field-position.ts
export enum FieldPosition {
    Pitcher = 'Pitcher',
    Catcher = 'Catcher',
    FirstBase = 'First Base',
    SecondBase = 'Second Base',
    ShortStop = 'Short Stop',
    ThirdBase = 'Third Base',
    RightField = 'Right Field',
    CenterRightField = 'Center Right Field',
    CenterField = 'Center Field',
    CenterLeftField = 'Center Left Field',
    LeftField = 'Left Field'
}